use dbl_max instead of get_float8_infinity()

+++ methods/array_ops/src/pg_gp/array_ops.c
@@ -525,7 +525,7 @@ array_min(PG_FUNCTION_ARGS){
 
     ArrayType *v = PG_GETARG_ARRAYTYPE_P(0);
     Oid element_type = ARR_ELEMTYPE(v);
-    Datum res = General_Array_to_Element(v, Float8GetDatum(0), FLT_MAX,
+    Datum res = General_Array_to_Element(v, Float8GetDatum(0), DBL_MAX,
             element_min, noop_finalize);
 
     PG_FREE_IF_COPY(v, 0);
@@ -544,7 +544,7 @@ array_max(PG_FUNCTION_ARGS){
     ArrayType *v = PG_GETARG_ARRAYTYPE_P(0);
     Oid element_type = ARR_ELEMTYPE(v);
 
-    Datum res = General_Array_to_Element(v, Float8GetDatum(0), -FLT_MAX,
+    Datum res = General_Array_to_Element(v, Float8GetDatum(0), -DBL_MAX,
             element_max, noop_finalize);
 
     PG_FREE_IF_COPY(v, 0);
@@ -573,7 +573,7 @@ array_max_index(PG_FUNCTION_ARGS) {
     }
 
     value_index *result = (value_index *)palloc(sizeof(value_index));
-    result->value = -FLT_MAX;
+    result->value = -DBL_MAX;
     result->index = 0;
 
     Datum res = General_Array_to_Struct(v, result, element_argmax, value_index_finalize);
@@ -598,7 +598,7 @@ array_min_index(PG_FUNCTION_ARGS) {
     }
 
     value_index *result = (value_index *)palloc(sizeof(value_index));
-    result->value = FLT_MAX;
+    result->value = DBL_MAX;
     result->index = 0;
 
     Datum res = General_Array_to_Struct(v, result, element_argmin, value_index_finalize);
